refactor(measurement): split model usage and group analysis
