+++ tools/measurement/measurement_tools/benchmark_analysis_math.py
@@ -19,14 +19,17 @@ def pipeline_stage_rows(dataframe: pd.DataFrame) -> pd.DataFrame:
         return stages.iloc[0:0]
     return stages[stages["stage"] == "Anonymizer._run_internal"]
 
+
 def first_float(frames: list[pd.DataFrame], columns: list[str]) -> float | None:
     value = first_value(frames, columns)
     return float(value) if value is not None else None
+
 
 def non_null_count(dataframe: pd.DataFrame, column: str) -> int:
     if column not in dataframe.columns:
         return 0
     return int(pd.to_numeric(dataframe[column], errors="coerce").notna().sum())
+
 
 def sum_optional_numbers(*values: object) -> float | None:
     numeric_values = [optional_number(value) for value in values]
@@ -34,24 +37,29 @@ def sum_optional_numbers(*values: object) -> float | None:
         return None
     return sum(cast(float, value) for value in numeric_values)
 
+
 def sum_bool_or_zero(dataframe: pd.DataFrame, column: str) -> int:
     if column not in dataframe.columns:
         return 0
     return int(dataframe[column].fillna(False).astype(bool).sum())
+
 
 def records_of_type(dataframe: pd.DataFrame, record_type: str) -> pd.DataFrame:
     if "record_type" not in dataframe.columns:
         return dataframe.iloc[0:0]
     return dataframe[dataframe["record_type"] == record_type]
 
+
 def first_int(frames: list[pd.DataFrame], columns: list[str]) -> int | None:
     value = first_value(frames, columns)
     return int(float(value)) if value is not None else None
+
 
 def optional_number(value: object) -> float | None:
     if value is None or pd.isna(value):
         return None
     return float(value)
+
 
 def safe_rate(numerator: object, elapsed_sec: object) -> float | None:
     numerator_value = optional_number(numerator)
@@ -59,6 +67,7 @@ def safe_rate(numerator: object, elapsed_sec: object) -> float | None:
     if numerator_value is None or elapsed_value is None or elapsed_value <= 0:
         return None
     return numerator_value / elapsed_value
+
 
 def first_value(frames: list[pd.DataFrame], columns: list[str]) -> str | None:
     for frame in frames:
@@ -70,10 +79,12 @@ def first_value(frames: list[pd.DataFrame], columns: list[str]) -> str | None:
                 return str(values.iloc[0])
     return None
 
+
 def f1(precision: float | None, recall: float | None) -> float | None:
     if precision is None or recall is None or precision + recall == 0:
         return None
     return 2 * precision * recall / (precision + recall)
+
 
 def zero_with_positive_count(dataframe: pd.DataFrame, *, zero_column: str, positive_column: str) -> int:
     if zero_column not in dataframe.columns or positive_column not in dataframe.columns:
@@ -82,11 +93,13 @@ def zero_with_positive_count(dataframe: pd.DataFrame, *, zero_column: str, posit
     positive_values = pd.to_numeric(dataframe[positive_column], errors="coerce")
     return int(((zero_values == 0) & (positive_values > 0)).sum())
 
+
 def coalesce_number(*values: float | None) -> float | None:
     for value in values:
         if value is not None:
             return value
     return None
+
 
 def coerce_count_mapping(value: object) -> dict[str, int]:
     payload = value
@@ -104,10 +117,12 @@ def coerce_count_mapping(value: object) -> dict[str, int]:
             counts[str(key)] = int(numeric.iloc[0])
     return counts
 
+
 def records_of_types(dataframe: pd.DataFrame, record_types: set[str]) -> pd.DataFrame:
     if "record_type" not in dataframe.columns:
         return dataframe.iloc[0:0]
     return dataframe[dataframe["record_type"].isin(record_types)]
+
 
 def request_failure_rate(*, failed: object, total: object) -> float | None:
     failed_value = optional_number(failed)
@@ -116,11 +131,13 @@ def request_failure_rate(*, failed: object, total: object) -> float | None:
         return None
     return failed_value / total_value
 
+
 def positive_count(dataframe: pd.DataFrame, column: str) -> int:
     if column not in dataframe.columns:
         return 0
     values = pd.to_numeric(dataframe[column], errors="coerce").fillna(0)
     return int((values > 0).sum())
+
 
 def safe_ratio(numerator: object, denominator: object) -> float | None:
     numerator_value = optional_number(numerator)
@@ -129,9 +146,11 @@ def safe_ratio(numerator: object, denominator: object) -> float | None:
         return None
     return numerator_value / denominator_value
 
+
 def sum_int_or_none(dataframe: pd.DataFrame, column: str) -> int | None:
     value = _sum_or_none(dataframe, column)
     return int(value) if value is not None else None
+
 
 def sum_prefixed_ints(dataframe: pd.DataFrame, prefix: str) -> dict[str, int]:
     totals: dict[str, int] = {}
@@ -146,14 +165,39 @@ def sum_prefixed_ints(dataframe: pd.DataFrame, prefix: str) -> dict[str, int]:
             totals[column.removeprefix(prefix)] = value
     return totals
 
+
 def zero_count(dataframe: pd.DataFrame, column: str) -> int:
     if column not in dataframe.columns:
         return 0
     values = pd.to_numeric(dataframe[column], errors="coerce").dropna()
     return int((values == 0).sum())
 
+
 def model_workflow_rows(dataframe: pd.DataFrame) -> pd.DataFrame:
     return records_of_types(dataframe, {"ndd_workflow", "model_workflow"})
 
 
-__all__ = ['coalesce_number', 'coerce_count_mapping', 'f1', 'first_float', 'first_int', 'first_value', 'model_workflow_rows', 'non_null_count', 'optional_number', 'pipeline_stage_rows', 'positive_count', 'records_of_type', 'records_of_types', 'request_failure_rate', 'safe_rate', 'safe_ratio', 'sum_bool_or_zero', 'sum_int_or_none', 'sum_optional_numbers', 'sum_prefixed_ints', 'zero_count', 'zero_with_positive_count']
+__all__ = [
+    "coalesce_number",
+    "coerce_count_mapping",
+    "f1",
+    "first_float",
+    "first_int",
+    "first_value",
+    "model_workflow_rows",
+    "non_null_count",
+    "optional_number",
+    "pipeline_stage_rows",
+    "positive_count",
+    "records_of_type",
+    "records_of_types",
+    "request_failure_rate",
+    "safe_rate",
+    "safe_ratio",
+    "sum_bool_or_zero",
+    "sum_int_or_none",
+    "sum_optional_numbers",
+    "sum_prefixed_ints",
+    "zero_count",
+    "zero_with_positive_count",
+]
